migrate min and max of Field from long to double

+++ core/src/main/java/xyz/v2my/easymodeling/randomizer/NumberRandomizer.java
@@ -2,5 +2,5 @@ package xyz.v2my.easymodeling.randomizer;
 
 public abstract class NumberRandomizer<T> extends GenericRandomizer<T> {
 
-    public abstract T next(long min, long max);
+    public abstract T next(double min, double max);
 }
