+++ core/src/main/java/xyz/v2my/easymodeling/randomizer/numberrandomizer/FloatRandomizer.java
@@ -6,7 +6,7 @@ import xyz.v2my.easymodeling.randomizer.NumberRandomizer;
 public class FloatRandomizer extends NumberRandomizer<Float> {
 
     @Override
-    public Float next(long min, long max) {
+    public Float next(double min, double max) {
         return doubleBetween(min, max).floatValue();
     }
 }
