+++ core/src/main/java/xyz/v2my/easymodeling/randomizer/numberrandomizer/LongRandomizer.java
@@ -6,7 +6,7 @@ import xyz.v2my.easymodeling.randomizer.NumberRandomizer;
 public class LongRandomizer extends NumberRandomizer<Long> {
 
     @Override
-    public Long next(long min, long max) {
+    public Long next(double min, double max) {
         return doubleBetween(min, max).longValue();
     }
 }
